## Change Summary

Tracking when default arguments are override and allows a `skip_defaults` kwarg in the `dict` method. 

A problem that I have not yet solved is that a BaseModel goes through the `dict_validator` and is then passed back into `BaseModel.validate` after being cast to a dict, this does not allow the `__fields_set__` to come through at the moment and recursive models do not work. Any suggestions here?

```
make benchmark-pydantic
Before:           pydantic best=29.726μs/iter avg=30.439μs/iter stdev=0.563μs/iter
After:            pydantic best=29.061μs/iter avg=29.695μs/iter stdev=0.473μs/iter
```
## Related issue number

Discussion in #378.

## Checklist

* [x] Unit tests for the changes exist
* [ ] Tests pass on CI and coverage remains at 100%
* [ ] Documentation reflects the changes where applicable
* [ ] `HISTORY.rst` has been updated
  * if this is the first change since a release, please add a new section
  * include the issue number or this pull request number `#<number>`
  * include your github username `@<whomever>`