@samuelcolvin 2019-02-10
Single quotes please everywhere, also you'll need to resplect black. Use `make format`.