@dgasmith 2019-02-11
Of course, I was holding off on that area until we had an idea of what to do with the recursive copy operation. Easiest thing would be to write a custom copy validator for BaseModels rather than relying on the standard `dict` validator. However, that solution isn't as elegant as the current code.