@dgasmith 2019-02-11
Similar to #265. If you are fine with having a custom validator for BaseModels rather than the `dict_validator` we could solve both the issue here and #265.