@dgasmith 2019-02-11
#265 was deeper than I expected. I think that can be solved however, but I will leave it for another PR. This is now ready to review.