@dgasmith 2019-02-12
Odd on timings, those were 3.6. For 3.7 I see:
```
this branch               pydantic best=6.871μs/iter avg=7.165μs/iter stdev=0.169μs/iter
master                    pydantic best=6.841μs/iter avg=7.238μs/iter stdev=0.170μs/iter
```

Agreed on the keys comment, patched.