@samuelcolvin 2019-02-13
For me this is now ready to be merged. Please check my changes and let me know if you disagree with anything or have any questions.

We might need to update history to state that `contruct`'s signature has changed.