@dgasmith 2019-02-13
Looks good! I was hesitant to change `constructor`, but this is much better and is extensible if there are other slots needed in the future. Thanks for working through this and the great feedback.